Create new tunnel using the old as a reference
Create the ExternalVPNGateway
==

This patch creates the ExternalVPNGateway resource which is required to
create the VPN Tunnel.  The ExternalVPNGateway references the dynamic IP
address of the peer.

Creation of the ExternalVPNGateway is an asynchronous Operation.  This
patch implements waitDone() using go routines and channels to poll for
the completion of the operation using [globalOperations.get][get].
[globalOperations.wait][wait] is a future optimization.

[get]: https://cloud.google.com/compute/docs/reference/rest/v1/globalOperations/get
[wait]: https://cloud.google.com/compute/docs/reference/rest/v1/globalOperations/wait