Add dns update command to update DNS records
This patch adds a `scarab dns update` command to update records.  The
intent is to pass in a new IP address when a new dynamic IP is obtained
and automatically update at least the apex and the wildcard A records in
Cloud DNS.

Example usage:

    ./bin/scarab dns update --dnszone=jeff --rrnames="jeff.ois.run.,*.jeff.ois.run." --rrdatas=75.164.109.241

If rrdatas is not provide, the record type, name is scheduled for
deletion.  Note, if the record is already at the same rrdatas value then
the change is still submitted, though this isn't for any intetional
purpose.